Modify the backend spec to allow for fixing #17.
Allow teachers to change the team a user is in to any team they are a
member of already, as long as that user is not a BlueShirt.

BlueShirts can be in several teams, but this interface only allows
for the setting of a single team, which could potentially have bad
side effects if applied to BlueShirts.

Also expose the team that a user is in as part of the /user/:username
endpoint, so that the front-end can both display this information, as
well as a collection of suitable teams for changing to.

Change-Id: I4f9e11b0d37973e6fc8d1e138efeff93b0d3876e