Add backend support for email address validation.
This modifies how POST /user/:userid works, so that the request is
stored in the local sqlite database, with LDAP unchanged in regards
to the email address. Such requests can be cancelled by another POST
which contains the original address. These requests also email the
new address asking for confirmation that it exists.

Add a new GET /verify/:email:code endpoint, which is deliberately
not documented as part of the API. This is the target used in the
emails sent to the new address, and it is this endpoint which, when
given the right verification code, will actually modify LDAP with
the users new email address.

Also bump libnemesis to use a new property for the users' first name.

Note that, for the moment, this only applies to changes to existing
users' email addresses.

Change-Id: Ibaf5f2913e69a374c63f3c1b3e9a7967a2d916f0