Avoid claiming the current email is also new.
Detect that the current email address is the same as the one in the
current request for a change, and don't put out the new_email key in
such cases. This is actually a valid state to be in - just after a
change has been completed, the request is left in the database.

I've decided to leave the requests in place to cope with the user
clicking the link several times -- this ensures that they just get
the success message several times.

Change-Id: I7103aee14721df243a3bb5c2e3a2624e492e6dc5