Have users directly set a new password on the activation page.
Replace the static page which told users to go somewhere else to
change their password with a dynamic page that feels much more like
the rest of usreman. This page contains a password setting form that
ensures the password is long enough (6 chars) and matches a confirmation
field. After submission, the user is then shown a similar message
which confirms their details and links them to the forums etc.

Change-Id: I79254cbd0752ab67c128fa916643d35edb3e1aeb