Save emails to the db rather than sending immediately.
Introduce functionality to save requested emails into the sqlite
database rather than sending them directly. This should give us
greater control over when & how they're sent, as well as giving us
a record of what emails are sent. Add a config option which allows
for direct sending if desired (though the database will still be
populated).

Add tests for the mailer, particularly the new functionality, but
nothing as yet for actually sending the emails.

Have cron make use of a new helper function which sends the next
50 emails from the database, though we may want a specific script
that can be run more often.

Change-Id: I7efc93e68547313f5cfaa16d43288a811116457b