Prevent non-students accidentally removing themselves from teams.
At the moment, the backend API only supports the idea of setting a
user into a single team. While this is a valid assumption for students,
it's not valid for team-leaders or blueshirts. Previously, this
assumption (and limitation) were also present in the web UI as well.

The result of this is that if a non-student submitted their own
edit form, they would be moved into _only_ whichever team it showed
them in.

This commit prevents either non-student user type modifying their
own teams (not something they should need to do anyway), while also
changing the UI to show the possibly multiple teams they're in.
Eventually, it may be needed to let these users modify their own
teams, and the way these are now presented is compatible with that.

Change-Id: Ie1894092e16ee12aa6e781a43634d45eebbbedb9