Move registration failure handling next to success handling.
This gets both bits of functionality in the same place, making it
much clearer to see how they work. No functionality changes yet,
but this already makes it easy to spot the errors in this code.
The send_registration_hash method is now back to being a utility,
more useful now that it also wraps a failure callback too.

Change-Id: Ic9014c4af7581fe8646a5ccb76d8a5870b21bec9