Fix registration submission in case of partial failure.
Previously, if all of the submitted users worked, but some of the
rows had failed at the javascript check, then when the submissions
returned, the page was closed anyway because the pre-failures were
overlooked. This commit re-works the success criteria by adding a
second count to the registrations collection, which indicates the
total number of rows being worked on. This is then used to work out
if the page worked entirely.

To avoid confusing the user with submissions that worked subsequently
failing, these are now removed as soon as they return from the server,
leaving only those that failed (from either step).

Change-Id: I3d4a5d839d3a88f80eeca613d9be1418a0f1f6f1