Fix #2170: Prevent Students changing their names.
Add support in the backend for ignoring attempts to change the name
of a user by a student (which covers the case of student self-edits,
and any other cases of edits by a student aren't allowed anyway) as
well as a note in the spec about this. It follows the existing pattern
of just ignoring these attempts, rather than actively failing, so
that other changes are allowed.

While this is non-ideal, it will do for now as this commit also
modifies the front-end to disable the name fields when students are
logged in.

Change-Id: I6b0e091f37067f38dd8fc52faf3e49905fd37735