Reject registrations of names that don't start with a letter.
Since we use the names to construct the username, and then check
whether that name (plus an incrementing number) exists, there's the
potential for the user to either mess with our LDAP (unlikely with
the two characters available) or just get LDAP to cause us to 500
at them. The latter is much more likely, and results in the regisrations
form becoming a bit dead, which this should resolve.

Change-Id: Icf3e259a13b97d83c966d60f26c77a99b3adeaa8