Extract unit tests from the api tests, add test runner script.
This creates a new unit-tests directory under test/, which is used
to contain tests which test the nemesis internals directly. At some
point it may make sense for these internals and their tests to move
to libnemesis, but for now it's at least clearer what's what.

This also adds a ./run-tests script which will run all the current
tests (ie, not the client-tests) and return a summed return code.

Change-Id: I379e9e3ac7b9ff1277e540490de431d96064badb