Fix #2215: Fix self-modification in nemesis front-end.
At the moment, blueshirts don't actually have a type listing, so
neither of the radio buttons will be checked for them. Previously,
this resulted in the JS erroring because the .value attribute didn't
exist on the undefiend 0th index, and the form attempted to submit
via plain GET. Avoid this by checking the length of the returned
array, and only considering setting the value if it is suitable.

Change-Id: I6d7d3c527a57129a326992cd5651cdfee0687bbd