Update testing of user display, add test for the self-link.
This doesn't yet cover actual user editing, or any specifics, just
that the right box is shown, with the right username.

In order to do this, add helper methods which wait for a given element
to be shown, based around a helper method which waits for a predicate
or a timeout.

Change-Id: I76cc430b93b1f0f2f3c8519dd788e95b19a99e7d