Change default Chrome handling to go via the server, update readme.
This modifies the chrome browser example to allow either direct
control of the driver, which it turns out is slow, or access via the
server which is faster. Also update the readme to reflect this, and
tidy it up in the process now that we (almost) always need the server.

Change-Id: I9ab86a8aaf2cd9dbf6793f9ac287e0eaa34321a6