Merge and update the registration display and function tests.
There seems little point having one test which is going to effectively
duplicate the first portion of another test, as it may as well just
be one test. Ensure that the created user is not present in either
LDAP or sqlite either before or after the test runs so that it
should always work. Include handling of the feedback display so that
developers can see (via assertions) what error messages the user
would see.

Change-Id: I0d1159b0583ccf348e7f73d60e4eef32cd41b556