Add a tests for college refresh after adding/removing users.
This introduces a new user whose username doesn't begin with a number,
since numbers aren't valid first characters for html ids or classes
and would thus cause issues.
We ensure that the user is either added or removed to the list as
expected. In the remove case, check that if editing another user,
the edit session is unaffected, but if editing the removed user,
that the edit session is closed.

Change-Id: I8558f9a72ac8698e7c1b77c9c655fb1bcb25f0bf